requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.1
requests==2.26.0
boto3==1.18.65
gunicorn==20.1.0
openai-whisper==20230314